# fix: events: address sqlite index selection performance regressions